refactor(react): rewrite from esm to cjs
