+++ packages/eslint-config-react/src/eslint-plugin-jsx-a11y.js
@@ -1,4 +1,4 @@
-const rules = {
+module.exports = {
   /**
    * Enforce that a label tag has a text label and an associated control.
    * @see https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/main/docs/rules/label-has-associated-control.md
@@ -11,8 +11,4 @@ const rules = {
     assert: 'nesting',
     depth: 25,
   }],
-};
-
-export {
-  rules,
 };
